+++ internal/surfacers/file/file.go
@@ -24,10 +24,10 @@ import (
 	"sync"
 	"time"
 
+	"github.com/cloudprober/cloudprober/internal/surfacers/common/compress"
 	"github.com/cloudprober/cloudprober/logger"
 	"github.com/cloudprober/cloudprober/metrics"
-	"github.com/cloudprober/cloudprober/surfacers/common/compress"
-	"github.com/cloudprober/cloudprober/surfacers/common/options"
+	"github.com/cloudprober/cloudprober/surfacers/options"
 
 	configpb "github.com/cloudprober/cloudprober/internal/surfacers/file/proto"
 )
