+++ internal/surfacers/probestatus/probestatus_test.go
@@ -24,7 +24,7 @@ import (
 	configpb "github.com/cloudprober/cloudprober/internal/surfacers/probestatus/proto"
 	"github.com/cloudprober/cloudprober/metrics"
 	"github.com/cloudprober/cloudprober/state"
-	"github.com/cloudprober/cloudprober/surfacers/common/options"
+	"github.com/cloudprober/cloudprober/surfacers/options"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/proto"
 )
